Allow using a DB URL as session name.
In fact, any URL pointing to a GC3Pie `Store` will do.
Since, in this case, the session metadata is missing
(as it's normally only stored in the session directory)
then we reconstruct it or just guess.

This allows one to use GC3Utils for inspecting DB-based task
stores; e.g.::

    gstat --brief --session postgresql://user:passwd@db.example.org/